Wire fabSampleFour to secondActivity, pass String, rename activities
pass string extra to SecondActivity and toast it
